requirements.txt
beautifulsoup4==4.13.5
build==1.3.0
cachetools==5.5.2
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
colorama==0.4.6
configparser==7.2.0
cryptography==45.0.7
docutils==0.22.2
et_xmlfile==2.0.0
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
httplib2==0.22.0
id==1.5.0
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
keyring==25.6.0
markdown-it-py==4.0.0
mdurl==0.1.2
more-itertools==10.8.0
nh3==0.3.1
numpy==2.2.4
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
plyer==2.1.0
prettytable==3.16.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.30.2
pwinput==1.0.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
readme_renderer==44.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.2.0
rsa==4.9.1
SecretStorage==3.4.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
twine==6.2.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
wheel==0.45.1